+++ src/test/java/de/dm/intellij/liferay/schema/LiferayJournalStructureJsonSchemaDataDefinitionFileProviderTest.java
@@ -1,0 +1,30 @@
+package de.dm.intellij.liferay.schema;
+
+import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.testFramework.LightProjectDescriptor;
+import com.intellij.testFramework.fixtures.BasePlatformTestCase;
+import de.dm.intellij.test.helper.LightProjectDescriptorBuilder;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+
+public class LiferayJournalStructureJsonSchemaDataDefinitionFileProviderTest extends BasePlatformTestCase {
+
+    @NotNull
+    @Override
+    protected LightProjectDescriptor getProjectDescriptor() {
+        return LightProjectDescriptorBuilder.LIFERAY_SCHEMA_ROOT_ACCESS_PROJECT_DESCRIPTOR;
+    }
+
+    @Override
+    protected String getTestDataPath() {
+        return "testdata/de/dm/intellij/liferay/schema/LiferayJournalStructureJsonSchemaDataDefinitionFileProviderTest";
+    }
+
+    public void testCompletion() {
+        myFixture.configureByFiles("WEB-INF/src/resources-importer/journal/structures/test.json");
+        myFixture.complete(CompletionType.BASIC, 1);
+        List<String> strings = myFixture.getLookupElementStrings();
+        assertTrue(strings.contains("dataDefinitionFields"));
+    }
+}